Add executeShellRaw method. Avoids buffering issues
The buffering used in executeShell causes issues when using shell:logcat,
the character per character implementation of the InputStream handling in AdbFilterInputStream together with the BufferedInputStream class
cause that the caller only gets result once either the process ends or 8k of output is avilable. The executeShellRaw method bypasses
the custom buffer handling, allowing the caller to receive the output without delay.
Note that the issue is only apparent when using a long running commamd with little output.